% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{binom.pexplore}
\alias{binom.pexplore}
\title{Explore the power of a binomial experiment under different sample sizes using pilot data.}
\usage{
binom.pexplore(lown, topn, pilotdata, r = 10000,
  alternative = c("two.sided", "less", "greater"), alpha = 0.05,
  nullp = 0.5, conf.level = 0.95, plotit = TRUE)
}
\arguments{
\item{lown}{smallest sample size to explore.}

\item{topn}{largest sample size to explore.}

\item{r}{number of simulations to compute power.}

\item{alternative}{type of alternative hypothesis in binomial test. Must be "\code{two.sided}" (default), "\code{greater}", or "\code{less}".}

\item{alpha}{significance threshhold.}

\item{nullp}{probability of success in null hypothesis.}

\item{conf.level}{size of confidence intervals.}

\item{plotit}{logical (default=\code{TRUE}) value. Function generates a plot when \code{TRUE} and returns a data frame otherwise.}
}
\value{
The probability of finding \eqn{p < \alpha} with the experiment description and a 95% Clopper-Pearson confidence interval.
}
\description{
\code{binom.pexplore} computes (through simulation) the power of a binomial experiment under different sample sizes.
Rather than taking a probability of success (like \code{binom.explore}), \code{binom.pexplore} takes a vector of pilot data.
}
\examples{
binom.pexplore(lown=16, topn=24, pilotdata = (1, 1, 0, 0, 1, 1), p=0.8)
binom.pexplore(lown=16, topn=24, pilotdata = ("a", "b", "b", "b"), p=0.8, alternative="greater")
binom.pexplore(lown=16, topn=24, pilotdata = ("a", "b", "b", "b"), r=5000, nullp=0.25, alternative="greater")
}
\seealso{
\code{\link{binom.power}}, \code{\link{binom.ppow}}, \code{\link{binom.explore}}, and \code{\link{binom.pexplore}}.
}
