% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.R
\name{chisq.pow}
\alias{chisq.pow}
\title{Compute the power using a chi-squared test.}
\usage{
chisq.pow(x, y = NULL, n, r = 10000, alpha = 0.05, conf.level = 0.95)
}
\arguments{
\item{x}{a list containing probabilities of each category.}

\item{y}{an optional list with the probabilities of the second group.}

\item{n}{sample size.}

\item{r}{number of simulations to compute power.}

\item{alpha}{significance threshhold.}

\item{conf.level}{size of confidence intervals.}

\item{mu}{mean value according to null hypothesis (default = \code{0}). Only used in one sample t-tests.}
}
\value{
The probability of finding \eqn{p < \alpha} with the experiment description.
}
\description{
\code{chisq.pow} computes (via simulation) the power of an experiment that will be analyzed using a chi-square test.
}
\examples{
chisq.pow(x=c(0.2, 0.5, 0.3), n=32) # Chi-squared distribution against chance with 32 observations.
chisq.pow(x=c(0.2, 0.5, 0.3), y=c(0.8, 0.1, 0.1)) # Compare both group with 16 observations in each group.
}
\seealso{
\code{\link{chisq.pow}}, \code{\link{chisq.ppow}}, \code{\link{chisq.explore}}, and \code{\link{chisq.pexplore}}.
}
