% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishers.R
\name{fishers.pow}
\alias{fishers.pow}
\title{Compute the power using a one- or two-sample unpaired t-test.}
\usage{
fishers.pow(means, var, n, r = 10000, alternative = c("two.sided", "less",
  "greater"), mu = NULL, alpha = 0.05)
}
\arguments{
\item{means}{either a list with two average values (computes a two-sample t-test) or a single value (computes a one-sample t-test).}

\item{var}{expected variance in each group.}

\item{n}{sample size.}

\item{r}{number of simulations to compute power.}

\item{alternative}{type of alternative hypothesis in binomial test. Must be "\code{two.sided}" (default), "\code{greater}", or "\code{less}".}

\item{mu}{mean value according to null hypothesis (default = \code{0}). Only used in one sample t-tests.}

\item{alpha}{significance threshhold.}
}
\value{
The probability of finding \eqn{p < \alpha} with the experiment description.
}
\description{
\code{ttest.pow} computes (via simulation) the power of an experiment that will be analyzed using a t-test. When two means are provided,
function assumes a two-sample unpaired t-test, and \code{n} is interpreted as the sample size of each group (for a total sample size or \code{2n}).
}
\examples{
ttest.pow(means=c(5, 10), var=10, n=16) # two-sample t-test. n=16 refers to each condition, for a total of 32.
ttest.pow(means=20, var=10, n=16) # one-sample t-test. Comparing if average is different from 0. Because there is only condition, the total sample isze is 16.
ttest.pow(means=20, var=10, n=16, mu=10, alternative="higher") # one-sample t-test. Comparing if average is higher than 10.
}
\seealso{
\code{\link{ttest.pow}}, \code{\link{ttest.ppow}}, \code{\link{ttest.explore}}, and \code{\link{ttest.pexplore}}.
}
