% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm.explore}
\alias{lm.explore}
\title{Compute the power of a linear model for a range of sample sizes.}
\usage{
lm.explore(lmodel, lown, topn, r = 10000, alpha = 0.05, conf.level = 0.95,
  plotit = TRUE)
}
\arguments{
\item{lmodel}{linear model trained through lm object.}

\item{lown}{lowest sample size.}

\item{topn}{highest sample size.}

\item{r}{number of simulations to compute power.}

\item{alpha}{significance threshhold.}

\item{conf.level}{size of confidence intervals.}
}
\value{
The probability of finding \eqn{p < \alpha} with the experiment description and a 95\% Clopper-Pearson confidence interval for each item in the equation.
}
\description{
\code{lm.explore} computes (via simulation) the power of a linear model trained with pilot data.
}
\examples{
library(datasets)
Model <- lm(mpg ~ hp + vs, data = mtcars)
lm.pow(n=16, Model)
}
