% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm.pow}
\alias{lm.pow}
\title{Compute the power of a linear model.}
\usage{
lm.pow(lmodel, n, r = 10000, alpha = 0.05, conf.level = 0.95)
}
\arguments{
\item{lmodel}{linear model trained through lm object.}

\item{n}{sample size.}

\item{r}{number of simulations to compute power.}

\item{alpha}{significance threshhold.}

\item{conf.level}{size of confidence intervals.}
}
\value{
The probability of finding \eqn{p < \alpha} with the experiment description and a 95\% Clopper-Pearson confidence interval for each item in the equation.
}
\description{
\code{lm.pow} computes (via simulation) the power of a linear model trained with pilot data.
}
\examples{
library(datasets)
Model <- lm(mpg ~ hp + vs, data = mtcars)
lm.pow(n=16, Model)
}
