% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpa.R
\docType{package}
\name{mcpa}
\alias{mcpa}
\alias{mcpa-package}
\title{mcpa: Intuitive power analysis through simulation}
\description{
mcpa (monte carlo power analyses) is a package for power analyses. In contrast to classic power
analyses that rely on Cohen's delta, mcpa computes all powerby simulating the process of gathering
and analyzing data thousands of time.
}
\section{mcpa functions}{

}

