% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishers.R, R/ttest.R
\name{ttest.explore}
\alias{ttest.explore}
\alias{ttest.explore}
\title{Explore the power using a t-test under different sample sizes.}
\usage{
ttest.explore(lown, topn, means, var, r = 10000,
  alternative = c("two.sided", "less", "greater"), mu = 0, alpha = 0.05,
  conf.level = 0.95, plotit = TRUE)

ttest.explore(lown, topn, means, var, r = 10000,
  alternative = c("two.sided", "less", "greater"), mu = 0, alpha = 0.05,
  conf.level = 0.95, plotit = TRUE)
}
\arguments{
\item{lown}{smallest sample size to explore.}

\item{topn}{largest sample size to explore.}

\item{means}{either a list with two average values (computes a two-sample t-test) or a single value (computes a one-sample t-test).}

\item{var}{expected variance in each group.}

\item{r}{number of simulations to compute power.}

\item{alternative}{type of alternative hypothesis in binomial test. Must be "\code{two.sided}" (default), "\code{greater}", or "\code{less}".}

\item{mu}{mean value according to null hypothesis (default = \code{0}). Only used in one sample t-tests.}

\item{alpha}{significance threshhold.}

\item{plotit}{logical (default=\code{TRUE}) value. Function generates a plot when \code{TRUE} and returns a data frame otherwise.}

\item{lown}{smallest sample size to explore.}

\item{topn}{largest sample size to explore.}

\item{plotit}{logical (default=\code{TRUE}) value. Function generates a plot when \code{TRUE} and returns a data frame otherwise.}
}
\value{
The probability of finding \eqn{p < \alpha} with the experiment description.
}
\description{
\code{ttest.explore} computes (through simulation) the power of an experiment that will be analyzed using a t-test for a set of potential sample sizes. When two means are provided,
function assumes a two-sample unpaired t-test, and \code{n} is interpreted as the sample size of each group (for a total sample size or \code{2n}).

\code{ttest.explore} computes (through simulation) the power of an experiment that will be analyzed using a t-test for a set of potential sample sizes. When two means are provided,
function assumes a two-sample unpaired t-test, and \code{n} is interpreted as the sample size of each group (for a total sample size or \code{2n}).
}
\examples{
ttest.explore(lown=10, topn=15, means=c(5, 10), var=10) # two-sample t-test. Effective sample sizes are 20 to 30 (10 to 15 per group)
ttest.explore(lown=10, topn=15, means=20, var=10) # one-sample t-test. Comparing if average is different from 0.
ttest.explore(lown=10, topn=15, means=20, var=10, mu=10, alternative="higher") # one-sample t-test. Comparing if average is higher than 10.
ttest.explore(lown=10, topn=15, means=c(5, 10), var=10) # two-sample t-test. Effective sample sizes are 20 to 30 (10 to 15 per group)
ttest.explore(lown=10, topn=15, means=20, var=10) # one-sample t-test. Comparing if average is different from 0.
ttest.explore(lown=10, topn=15, means=20, var=10, mu=10, alternative="higher") # one-sample t-test. Comparing if average is higher than 10.
}
\seealso{
\code{\link{ttest.pow}}, \code{\link{ttest.ppow}}, \code{\link{ttest.explore}}, and \code{\link{ttest.pexplore}}.
}
