% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishers.R, R/ttest.R
\name{ttest.pexplore}
\alias{ttest.pexplore}
\alias{ttest.pexplore}
\title{Explore power as a function of sample sie using a one- or two-sample unpaired t-test by bootstrapping pilot data.}
\usage{
ttest.pexplore(x, y = NULL, lown, topn, r = 10000,
  alternative = c("two.sided", "less", "greater"), mu = NULL,
  alpha = 0.05, conf.level = 0.95, plotit = TRUE)

ttest.pexplore(x, y = NULL, lown, topn, r = 10000,
  alternative = c("two.sided", "less", "greater"), mu = NULL,
  alpha = 0.05, conf.level = 0.95, plotit = TRUE)
}
\arguments{
\item{lown}{smallest sample size to explore.}

\item{topn}{largest sample size to explore.}

\item{r}{number of simulations to compute power.}

\item{alternative}{type of alternative hypothesis in binomial test. Must be "\code{two.sided}" (default), "\code{greater}", or "\code{less}".}

\item{mu}{mean value according to null hypothesis (default = \code{0}). Only used in one sample t-tests.}

\item{alpha}{significance threshhold.}

\item{plotit}{logical (default=\code{TRUE}) value. Function generates a plot when \code{TRUE} and returns a data frame otherwise.}

\item{lown}{smallest sample size to explore.}

\item{topn}{largest sample size to explore.}

\item{plotit}{logical (default=\code{TRUE}) value. Function generates a plot when \code{TRUE} and returns a data frame otherwise.}
}
\value{
The probability of finding \eqn{p < \alpha} with the experiment description.

The probability of finding \eqn{p < \alpha} with the experiment description.
}
\description{
\code{ttest.pexplore} computes (via simulation) the power of an experiment that will be analyzed using a t-test for a range of sample sizes.
Rather than taking a theoretical distribution, this function takes empirical data and bootstraps them to calculate the power.
For an equivalent function that does not rely on pilot data see \link{ttestpower}.

\code{ttest.pexplore} computes (via simulation) the power of an experiment that will be analyzed using a t-test for a range of sample sizes.
Rather than taking a theoretical distribution, this function takes empirical data and bootstraps them to calculate the power.
For an equivalent function that does not rely on pilot data see \link{ttestpower}.
}
\examples{
ttest.pexplore(x=c(0, 5, 10), lown=16, topn=24) # Power for a one-sample t-test with n in 16-24. Pilot data consists of three data points.
ttest.pexplore(x=c(0, 5, 10), lown=16, topn=24,mu = -5) # Same as above, changing the avarege under the null to -5.
ttest.pexplore(x=c(0, 5, 10), lown=16, topn=24, y=c(9, 3, 2, 1)) # Power for a two-sample t-test with n=16-24 (per condition) using unbalanced pilot data.
ttest.pexplore(x=c(0, 5, 10), lown=16, topn=24) # Power for a one-sample t-test with n in 16-24. Pilot data consists of three data points.
ttest.pexplore(x=c(0, 5, 10), lown=16, topn=24,mu = -5) # Same as above, changing the avarege under the null to -5.
ttest.pexplore(x=c(0, 5, 10), lown=16, topn=24, y=c(9, 3, 2, 1)) # Power for a two-sample t-test with n=16-24 (per condition) using unbalanced pilot data.
}
\seealso{
\code{\link{ttest.pow}}, \code{\link{ttest.ppow}}, \code{\link{ttest.explore}}, and \code{\link{ttest.pexplore}}.
}
