% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishers.R, R/ttest.R
\name{ttest.ppow}
\alias{ttest.ppow}
\alias{ttest.ppow}
\title{Compute the power using a one- or two-sample unpaired t-test by bootstrapping pilot data.}
\usage{
ttest.ppow(x, y = NULL, n, r = 10000, alternative = c("two.sided", "less",
  "greater"), mu = NULL, alpha = 0.05, conf.level = 0.95)

ttest.ppow(x, y = NULL, n, r = 10000, alternative = c("two.sided", "less",
  "greater"), mu = NULL, alpha = 0.05, conf.level = 0.95)
}
\arguments{
\item{x}{a data frame with two columns, or a list with pilot data.}

\item{y}{a list with pilot data. When x is a list and y is not provided, a one-tailed t-test is used.}

\item{n}{sample size.}

\item{r}{number of simulations to compute power.}

\item{alternative}{type of alternative hypothesis in binomial test. Must be "\code{two.sided}" (default), "\code{greater}", or "\code{less}".}

\item{mu}{mean value according to null hypothesis (default = \code{0}). Only used in one sample t-tests.}

\item{alpha}{significance threshhold.}

\item{x}{a data frame with two columns, or a list with pilot data.}

\item{y}{a list with pilot data. When x is a list and y is not provided, a one-tailed t-test is used.}
}
\value{
The probability of finding \eqn{p < \alpha} with the experiment description.

The probability of finding \eqn{p < \alpha} with the experiment description.
}
\description{
\code{ttest.ppow} computes (via simulation) the power of an experiment that will be analyzed using a t-test.
Rather than taking a theoretical distribution, this function takes empirical data and bootstraps them to calculate the power.
For an equivalent function that does not rely on pilot data see \link{ttest.pow}.

\code{ttest.ppow} computes (via simulation) the power of an experiment that will be analyzed using a t-test.
Rather than taking a theoretical distribution, this function takes empirical data and bootstraps them to calculate the power.
For an equivalent function that does not rely on pilot data see \link{ttest.pow}.
}
\examples{
ttest.ppow(x=c(0, 5, 10), n=16) # Power for a one-sample t-test with n=16. Pilot data consists of three data points.
ttest.ppow(x=c(0, 5, 10), n=16, mu = -5) # Same as above, changing the avarege under the null to -5.
ttest.ppow(x=c(0, 5, 10), y=c(9, 3, 2, 1), n=30) # Power for a two-sample t-test with n=30 (per condition) using unbalanced pilot data.
ttest.ppow(x=c(0, 5, 10), n=16) # Power for a one-sample t-test with n=16. Pilot data consists of three data points.
ttest.ppow(x=c(0, 5, 10), n=16, mu = -5) # Same as above, changing the avarege under the null to -5.
ttest.ppow(x=c(0, 5, 10), y=c(9, 3, 2, 1), n=30) # Power for a two-sample t-test with n=30 (per condition) using unbalanced pilot data.
}
\seealso{
\code{\link{ttest.pow}}, \code{\link{ttest.ppow}}, \code{\link{ttest.explore}}, and \code{\link{ttest.pexplore}}.
}
